requirements.txt
groq>=0.4.1
python-dotenv>=1.0.0
behave>=1.2.6
pytest>=7.4.0
pytest-html>=3.2.0
allure-behave>=2.13.2
jinja2>=3.1.2
pyyaml>=6.0.1
requests>=2.31.0

# UI automation / discovery
playwright>=1.42.0

# Local LLM (TinyLlama + LoRA)
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece
protobuf

# Optional: For GPU acceleration (uncomment if using CUDA)
# bitsandbytes>=0.41.0  # For quantization
